Index of Julia Functions

SequentialGWAS.get_action — Method

Compares the variant to the 1000 GP information and returns an action to take together with a reason.

Potential actions are

DROP

KEEP

FLIP

Some particularly unexpected ACTION (REASON) are:

"KEEP (REVERSE-REF-ALT)"

"FLIP (COMPLEMENT-NOT-MATCHING-KGP)"

Because they mean the minor/major alleles are reversed in our dataset as compared to the reference KGP.

SequentialGWAS.kgp_unrelated_individuals — Method

Only keeps the first individual within each family and writes them to outfile.

SequentialGWAS.read_bim — Method

read_bim(file)

Columns Description from: https://www.cog-genomics.org/plink/1.9/formats#bim

Chromosome code (either an integer, or 'X'/'Y'/'XY'/'MT'; '0' indicates unknown) or name

Variant identifier

Position in morgans or centimorgans (safe to use dummy value of '0')

Base-pair coordinate (1-based; limited to 231-2)

Allele 1 (corresponding to clear bits in .bed; usually minor)

Allele 2 (corresponding to set bits in .bed; usually major)

SequentialGWAS.read_fam — Method

read_fam(file)

Columns Description from: https://www.cog-genomics.org/plink/1.9/formats#fam

Family ID ('FID')

Within-family ID ('IID'; cannot be '0')

Within-family ID of father ('0' if father isn't in dataset)

Within-family ID of mother ('0' if mother isn't in dataset)

Sex code ('1' = male, '2' = female, '0' = unknown)

Phenotype value ('1' = control, '2' = case, '-9'/'0'/non-numeric = missing data if case/control)

SequentialGWAS.read_map — Method

read_map(file)

Columns Description from: https://www.cog-genomics.org/plink/1.9/formats

Chromosome code. PLINK 1.9 also permits contig names here, but most older programs do not.

Variant identifier

Position in morgans or centimorgans (optional; also safe to use dummy value of '0')

Base-pair coordinate

SequentialGWAS.report_qc_effect — Method

report_qc_effect(input_prefix, output_prefix)

Report the effect of the QC filtering process on SNPs and samples.

SequentialGWAS.write_map — Method

write_map(file_prefix, array)

SequentialGWAS.write_release_samples_to_drop — Method

We drop duplicate individuals according to the following priority:

WGS > More Recent Array > Older Array

SequentialGWAS.OneTimeChecks.array_overlap — Method

array_overlap(manifest_file1, manifest_file2)

Takes two manifest files from Illumina genotyping arrays and returns the intersection of the SNP names.

SequentialGWAS.OneTimeChecks.identify_snps_to_flip — Method

identify_snps_to_flip(manifest_file)

According to this link, the RefStrand column in the manifest file corresponds to the standard designation for all eukaryotic organisms used by HapMap and 1000 Genomes Project. Variants with RefStrand equal to - need to be flipped to the + strand.

SequentialGWAS.OneTimeChecks.make_snps_to_flip_list — Method

make_snps_to_flip_list(output, manifest_file)

Takes a manifest file from an Illumina genotyping array and writes a list of SNPs that are on the - strand and need to be flipped to the + strand.